requirements.txt
streamlit
torch
datasets
torchvision
pillow
tqdm
matplotlib